fused-chatbot/server/rag_service/requirements.txt
flask==3.0.2
langchain==0.3.19
langchain-community==0.3.38
langchain-core==0.3.43
faiss-cpu==1.7.4
sentence-transformers==2.5.1
pypdf==4.1.0
python-docx==1.1.0
python-pptx==0.6.23
google-generativeai==0.3.2
python-dotenv==1.0.1
gunicorn==21.2.0 